Move all of the native methods to Squirrel.dll
