+++ src/Update/Update-Mono.csproj
@@ -75,7 +75,6 @@
     <Compile Include="AnimatedGifWindow.cs" />
     <Compile Include="CopStache.cs" />
     <Compile Include="Mono.Options\Options.cs" />
-    <Compile Include="NativeMethods.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
